+++ src/gufo/thor/services/auth.py
@@ -12,7 +12,7 @@ Attributes:
 
 # Gufo Thor modules
 from .envoy import envoy
-from .liftbridge import liftbridge
+from .kafka import kafka
 from .migrate import migrate
 from .mongo import mongo
 from .noc import NocHcService
@@ -22,7 +22,7 @@ class AuthService(NocHcService):
     """auth service."""
 
     name = "auth"
-    dependencies = (envoy, liftbridge, migrate, mongo)
+    dependencies = (envoy, kafka, migrate, mongo)
     allow_scale = True
     expose_http_prefix = "/api/auth/"
     compose_command = (